# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FEMAQUA API",
        "description": "Documentação da API FEMAQUA para gerenciamento de repositório de ferramentas. A API oferece funcionalidades de criação, visualização, edição e exclusão de ferramentas, além de autenticação de usuários via tokens.",
        "contact": {
            "email": "willame.dev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tools": {
            "get": {
                "tags": [
                    "Tools"
                ],
                "summary": "Listar ferramentas",
                "operationId": "getTools",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filtrar ferramentas por tag",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "PHP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ferramentas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ToolResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tools"
                ],
                "summary": "Criar uma nova ferramenta",
                "operationId": "createTool",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "link",
                                    "description",
                                    "tags"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://laravel.com"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Framework PHP para aplicações web"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "PHP"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ferramenta criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tools/{tool}": {
            "get": {
                "tags": [
                    "Tools"
                ],
                "summary": "Obter uma ferramenta específica",
                "operationId": "getToolById",
                "parameters": [
                    {
                        "name": "tool",
                        "in": "path",
                        "description": "ID da ferramenta",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da ferramenta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ferramenta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tool not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tools"
                ],
                "summary": "Atualizar uma ferramenta existente",
                "operationId": "updateTool",
                "parameters": [
                    {
                        "name": "tool",
                        "in": "path",
                        "description": "ID da ferramenta",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "link",
                                    "description",
                                    "tags"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laravel"
                                    },
                                    "link": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://laravel.com"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Framework PHP atualizado"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "PHP"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ferramenta atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ToolResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permissão negada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not own this tool."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ferramenta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tool not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tools"
                ],
                "summary": "Excluir uma ferramenta",
                "operationId": "deleteTool",
                "parameters": [
                    {
                        "name": "tool",
                        "in": "path",
                        "description": "ID da ferramenta",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ferramenta excluída com sucesso"
                    },
                    "403": {
                        "description": "Permissão negada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You do not own this tool."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ferramenta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tool not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ToolResource": {
                "title": "Tool Resource",
                "description": "Detalhes da ferramenta",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Laravel"
                    },
                    "link": {
                        "type": "string",
                        "format": "url",
                        "example": "https://laravel.com"
                    },
                    "description": {
                        "type": "string",
                        "example": "Framework PHP para aplicações web"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "PHP"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via token Bearer. Insira o token no formato 'Bearer {token}'.",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}